more stuff, some untested
