+++ utils/listutils.py
@@ -28,6 +28,9 @@ def create_list(x, y):
         return x.append(y) or x
 
 
+
+
+
 def preobrazovatel(list_a, levels, level2=1):
     list_tmp = []
     level = levels[0]
